refactor[logging]: move type defs to the types folder
